represent xp as integer

+++ src/main/kotlin/bot/procyon/models/User.kt
@@ -13,6 +13,6 @@ data class User(
     @KomapperId
     val id:             Snowflake,
     val balance:        Int,
-    val exp:            Float
+    val exp:            Int
 )
